Make the `HTTPClient` a protocol
We can now inject the client to the `RemoteFeedLoader` so it no longer has the responsibility
of locating an object to call. Instead it just uses what was given to it.

As the client is now injected to the `RemoteFeedLoader`, we don't need the `shared` instance
anymore.

This change is reflected in the tests.